refactor: move helper files to aws folder
