+++ integrations/aws/aws/aws_credentials.py
